+++ src/main/java/com/marginallyclever/communications/presentation/CANOpen.java
@@ -1,8 +1,8 @@
 package com.marginallyclever.communications.presentation;
 
-import com.marginallyclever.communications.SessionLayer;
-import com.marginallyclever.communications.SessionLayerEvent;
-import com.marginallyclever.communications.SessionLayerListener;
+import com.marginallyclever.communications.session.SessionLayer;
+import com.marginallyclever.communications.session.SessionLayerEvent;
+import com.marginallyclever.communications.session.SessionLayerListener;
 import javax.swing.*;
 
 public class CANOpen implements SessionLayerListener {
