+++ src/main/java/com/marginallyclever/communications/serial/SerialSession.java
@@ -1,10 +1,8 @@
 package com.marginallyclever.communications.serial;
 
-import com.marginallyclever.communications.SessionLayer;
-import com.marginallyclever.communications.SessionLayerEvent;
+import com.marginallyclever.communications.session.SessionLayer;
+import com.marginallyclever.communications.session.SessionLayerEvent;
 import com.marginallyclever.communications.TransportLayer;
-import com.marginallyclever.convenience.log.Log;
-import com.marginallyclever.robotoverlord.RobotOverlord;
 import jssc.SerialPort;
 import jssc.SerialPortEvent;
 import jssc.SerialPortEventListener;
